Fix bug in manual column move
Repro for this bug:

1. Enable manualColumnMove
2. Click on the move tab of a column, but don't drag it

Expected: column is not moved

Actual: column can be moved to an arbitrary location

If this is the first time that move-column has been invoked since the
table was initialized, the afterColumnMove callback is actually invoked
with a newIndex parameter of undefined or NaN.

The cause is that endCol is not initialized unless the move tab is
dragged into a different cell. The simple fix is to always initialize
endCol = startCol as soon as we start dragging.