Bugfix for a previous bugfix in manual column move
If you clicked the move tab on a column while you were scrolled over to
the right, you'd get a bad move and a potentially illegal one. The
reason is that the colOffset was being added in to the endCol twice.
This only shows up when you click and *don't* move the column, because
when you *do* move the column, endCol gets computed again in a
different place, according to a correct formula.